Minor refactoring of Cache and constants
